driver: interrupt_controller: intc_xlnx: add Xilinx driver
Add driver for Xilinx interrupt controller. Controller can be used in
connection to Microblaze, Microblaze V or Xilinx ARM SOCs.
In Microblaze case it is primary interrupt controller.
In other cases it is secondary interrupt controller in cascade mode.

Patch is adding support for configuration for Microblaze V that's why
xilinx_intc_init() is checking if RISCV_IRQ_MEXT is present and creating
connection. And because it is secondary interrupt controller
CONFIG_2ND_LVL_ISR_TBL_OFFSET is passed via argument.

Co-developed-by: Alp Sayin <alpsayin@gmail.com>
Signed-off-by: Alp Sayin <alpsayin@gmail.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>