samples: video: capture: rt10xx: Add work-around for init priority issue
Similar to the mcxn947 (ov7670/smartDMA) issue, the CSI needs to be
initialized BEFORE the camera sensor to provide clock to the camera
sensor. However, the CSI node refers to the camera sensor node in the DT
so it has to be initialized AFTER the sensor to be able to compiled.

This is a deadlock issue as Zephyr currently does not allow circular
dependency. Disable the init priority check as a work-around for this.

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>