boards: opta: ethernet reorganization
This set of changes reorganize the ethernet configuration to avoid always
committing 16KB of RAM for the network buffers. The `status = "okay"` is
removed from ethernet configuration in the device tree and instead of
using a regulator to enable the PHY the correct GPIO pin is set only if
the network has been configured via CONFIG_NET_L2_ETHERNET.

As shown in the new overlay for the telnet sample it is sufficient to add
`status = "okay"` back to get a fully working network connection.

Signed-off-by: Federico Di Gregorio <fog@dndg.it>