drivers: video: move video-controls.h to video/controls.h
Move <zephyr/drivers/video-controls.h> to a new directory
<zephyr/drivers/video/controls.h>. The video-controls.h header was one
of the few exception of a driver having a separate header, with all
the other drivers having a directory to store all the extra headers
they would need. This also allow MAINTAINERS.yml to point at the whole
<zephyr/drivers/video/*> instead of individual files. The changes are
propagated through the tree. The original <zephyr/drivers/video.h>
itself header is not moved.

Signed-off-by: Josuah Demangeon <me@josuah.net>